test(throttle): cambio de namespaces para los endpoints throttle de la API
