+++ app/backend/tests/Feature/Api/V1/ThrottleOrderCreateTest.php
@@ -2,28 +2,31 @@
 
 declare(strict_types=1);
 
-namespace Tests\Feature;
+namespace Tests\Feature\Api\V1;
 
 use App\Models\User;
 use Illuminate\Foundation\Testing\RefreshDatabase;
 use Laravel\Sanctum\Sanctum;
 use Tests\TestCase;
 
-class ThrottleAuthenticatedTest extends TestCase
+class ThrottleOrderCreateTest extends TestCase
 {
     use RefreshDatabase;
 
     /**
-     * Test authenticated user throttle (should return 429 after 100 requests/min).
+     * Test order creation endpoint throttling (should return 429 after 100 requests/min).
      */
-    public function test_authenticated_throttle_returns_429_after_limit(): void
+    public function test_order_create_throttle_returns_429_after_limit(): void
     {
         $user = User::factory()->create();
         Sanctum::actingAs($user);
+        $payload = [
+            // Completa con los campos mínimos requeridos por Order
+        ];
         for ($i = 0; $i < 100; $i++) {
-            $this->getJson('/api/v1/me');
+            $this->postJson('/api/v1/orders', $payload);
         }
-        $response = $this->getJson('/api/v1/me');
+        $response = $this->postJson('/api/v1/orders', $payload);
         $response->assertStatus(429)
             ->assertJson([
                 'status' => false,
